All relevant details improved, ready for use for volume up and down
