+++ src/GestureControl/run_program.py
@@ -1,0 +1,4 @@
+from test import GestureDetector
+
+classifier = GestureDetector()
+classifier.run_backend_mainloop()
