+++ src/GestureControl/todos.txt
@@ -1,3 +1,3 @@
-1. Fix fingers detection so thumbs down gets triggered more often instead of being mistaken as multiple fingers
-2. Remove all other unnessesary things for now, leave palm close and open for future
-3. Segmentate different classes and files
+1. Fix fingers detection so thumbs down gets triggered more often instead of being mistaken as multiple fingers [done]
+2. Remove all other unnessesary things for now, leave palm close and open for future [done]
+3. Segmentate different classes and files [done]
